splitting files done; cleaned unused files
