--- scratch.py
@@ -1,5 +1,0 @@
-from main import WordClock
-from scratch2 import function1
-
-#instanceWC = WordClock()
-function1()
