--- scratch2.py
@@ -1,4 +1,0 @@
-def function1():
-    print("this is function1")
-
-function1()
