+++ wordClockDictionaries.py
@@ -1,0 +1,49 @@
+#Setup Dictionaries for different word groups
+dicIntro  = {
+  "1": "FRAU BABY",
+  "2": "FRAU BABY BERND",
+  "3": "BABY BERND",
+  "4": "MEIN BRAUNLING",
+  "5": "HI BRAUNBERT",
+  "6": "MEINE KATJA",
+  "7": "MEINE SONNE",
+  "8": "HI SCHATZ",
+}
+dicItIs  =  {
+  "1": "ES IST",
+  "2": "WIR HABEN",
+}
+dicMinute  =    {
+  "0": "UHR",
+  "1": "FUENF NACH",
+  "2": "ZEHN NACH",
+  "3": "VIERTEL NACH",
+  "4": "ZWANZIG NACH",
+  "5": "FUENF VOR HALB",
+  "6": "HALB",
+  "7": "FUENF NACH HALB",
+  "8": "ZEHN NACH HALB",
+  "9": "DREIVIERTEL",
+  "10": "ZEHN VOR",
+  "11": "FUENF VOR"
+}
+dicHour  =  {
+  "1": "EINS",
+  "2": "ZWEI",
+  "3": "DREI",
+  "4": "VIER",
+  "5": "FUENF",
+  "6": "SECHS",
+  "7": "SIEBEN",
+  "8": "ACHT",
+  "9": "NEUN",
+  "10": "ZEHN",
+  "11": "ELF",
+  "12": "ZWOELF",
+}
+dicEnd  =   {
+  "1": "KUSS",
+  "2": "KISS YOU",
+  "3": "MISS YOU",
+  "4": "LOVE YOU",
+}
